requirements.txt
flask
flask_sqlalchemy
flask_migrate
psycopg2
flask_expects_json
sqlacodegen